+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 
 - Made Vector::copyDataTo able to copy from device to host and vice versa
 
+- Added `diagSolve`, `max`, and `abs` vector operations.
+
 ## Changes to Re::Solve in release 0.99.2
 
 ### Major Features